third_party/WebKit/LayoutTests/platform/win/fast/dom/Window/Location/window-shadow-location-using-js-object-with-toString-expected.txt

PASS result is 'pass.html'
PASS successfullyParsed is true

TEST COMPLETE
